Wine Market
This application allows a customer process sales and returns on wines
which the user enters into the application.

When the ‘Process Sale’ button is pressed after entering all the
details, the program displays to the user the amount of the transaction
and their remaining current balance.

When the ‘Process Return’ is pressed, a refund amount is calculated
with 20% service charge added on. The current balance is also updated
to reflect the changes.